+++ packages/core/src/components/Bounds/__tests__/CenterHandle.test.tsx
