+++ packages/core/src/components/Bounds/__tests__/RotateHandle.test.tsx
