+++ packages/core/src/test/ContextWrapper.tsx
@@ -4,7 +4,6 @@ import { mockDocument } from './mockDocument'
 import { mockUtils } from './mockUtils'
 import { useTLTheme, TLContext, TLContextType } from '../hooks'
 import { Inputs } from '~inputs'
-import type { TLShape } from '~index'
 import type { BoxShape } from '~TLShapeUtil/TLShapeUtil.spec'
 
 export const ContextWrapper: React.FC = ({ children }) => {
