+++ packages/core/src/test/renderWithContext.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { render } from '@testing-library/react'
-import { ContextWrapper } from './context-wrapper'
+import { ContextWrapper } from './ContextWrapper'
 
 export const renderWithContext = (children: JSX.Element) => {
   return render(<ContextWrapper>{children}</ContextWrapper>)
